requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
openpyxl==3.1.2
python-dateutil==2.8.2
